refactor: rename to RouteNumberContainer
